Adds temp and tempcontrol modules.
The heater previously did a feedback loop, but we move that to tempcontrol.
Temp will handle the conversion from ADC values to temperature.

We keep using a single-byte set-value (tempc_value) so the EEPROM
handling in frontpanel doesn't have to know about 16-bits. 2*C
resolution is more than enough anyway.