Clamps and tunes tempcontrol.
Both the error and output is now clamped, allowing a higher Kp and
faster ramp-up. Clamping the error also means allowing more bits for
coefficients and accumulation.

The temp running mean isn't as important now that the ADC is doing
multisampling. It should probably be removed, and replaced with a
display update rate limiter.

Ki and Kd not yet tuned.