Fixes the ADC.
This has been bothering me for years, and been bad the past couple of
days while attempting to tune tempcontrol.

The default ADC prescaler is too fast for the 20 MHz the AT860D runs
at. This caused weird oscillating patterns where it really looked like
the ADC was skipping checking some bits (a wavy pattern if listing all
values and looking at the base-2 numbers.)

The potentiometer is just fine.

Also fixes divp2s16lf to set carry correctly if the divisor is
2^0. This is useful with add16cf.