Initial tuning of tempcontrol.
Makes pid clamp its accumulator. It's also the PID output, so clamping
it to the triac_zcc range makes sense.

Resets the pid accumulator when the set-temperature changes. It
accumulates negatively while at lowest-temp, which is slightly
annoying. Ideally, we'd have no I-term, since this is a pretty
straight forward system. It's just a bit slow.

Documented the tuning process.