Adds hysteresis to knob.
With all value ranges being smaller than the ADC resolution, it barely
shows.

One problem now is that the set-temperature changes when booting (with
DISABLE_SELFTEST,) because adc_knob_value reads zero until the ADC has
had a chance to run. This is similar to the temperature ADC channel
having a capacitor that needs charging before it shows a reasonable
value. The right solution is using the selftest mode, but a startup
timer would suffice.