Adds a calibration mode.
This is not like the original AT860D firmware. I understand the
intended split between "technician" (CAL), "supervisor" (SET) and
"worker", but the supervisor shouldn't update calibration values.

Based on a hunch, I suspect AT1 is setting the temp coefficient, while
AT2 sets the offset constant. Retaining the names, but changing the UX.