Fixes standby twitching at the temperature threshold.
The temperature goes up when disabling the airpump, so we use a latch
to keep the device in sleep. Another approach could have been a
hysteresis, but this is simpler, and the temperature is meant to be
low enough to be safe even with a few extra degrees. ("Safe" means:
doesn't set things on fire immediately, or burns out the heating
element.)